rename (components) to components
